requirements.txt
pandas>=2.2
numpy>=1.26
streamlit>=1.33
plotly>=5.20
scikit-learn>=1.3
PyYAML>=6.0
pyarrow>=15.0
python-dateutil>=2.9